Update README.md
Remote file downloading refers to the process of fetching a file from a remote server or location and saving it onto a local machine or device. This is a common task in many applications and can be used to retrieve data, documents, images, or any other type of file that is stored on a remote server.

To download a remote file, you typically need to know its URL or web address. This URL can be entered into a web browser or used in a program or script that supports remote file downloading. When the download is initiated, the file is transferred over the internet from the remote server to the local machine.

Remote file downloading can be performed using a variety of protocols, such as HTTP, FTP, or SFTP. Each protocol has its own set of advantages and disadvantages, and the choice of protocol will depend on the specific use case and requirements.

Overall, remote file downloading is a powerful and useful tool that enables users to access and retrieve files from remote servers quickly and easily.